+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
   [is planned](https://github.com/sleepingkingstudios/sleeping_king_studios-docs/issues/53)
 - [#25](https://gitlab.com/experimentslabs/federails/-/issues/25) - `Actor#local?` now resolves with a new `local` flag 
   on `Actor`, so it is now reliable.
+- Distant actors can now have local entities. Override `create_federails_actor_as_local?` in your models to determine
+  if associated actor is local or not (defaults to `true`)
 
 ### Maintenance